Pulled out Solana Utils
