+++ libs/src/services/solanaWeb3Manager/transfer.js
@@ -5,7 +5,7 @@ const {
 } = require('@solana/web3.js')
 const BN = require('bn.js')
 const borsh = require('borsh')
-const keccak256 = require('keccak256')
+const SolanaUtils = require('./utils')
 const secp256k1 = require('secp256k1')
 
 /**
@@ -69,17 +69,9 @@ async function transferWAudioBalance ({
     ...new BN(strippedEthAddress, 'hex').toArray('be')
   )
 
-  const ethPrivateKeyArr = Buffer.from(senderEthPrivateKey, 'hex')
-
   const senderSolanaPubkey = new PublicKey(senderSolanaAddress)
   const recipientPubkey = new PublicKey(recipientSolanaAddress)
-
-  // Hash the recipient solana pubkey and create signature
-  const msgHash = keccak256(recipientPubkey.toBytes())
-  const signatureObj = secp256k1.ecdsaSign(
-    Uint8Array.from(msgHash),
-    ethPrivateKeyArr
-  )
+  const { signature, recoveryId } = SolanaUtils.signBytes(recipientPubkey.toBytes(), senderEthPrivateKey)
 
   const instructionData = new TransferInstructionData({
     ethAddress: ethAddressArr,
@@ -135,8 +127,8 @@ async function transferWAudioBalance ({
   const secpTransactionInstruction = Secp256k1Program.createInstructionWithPublicKey({
     publicKey: Buffer.from(ethPubkey),
     message: recipientPubkey.toBytes(),
-    signature: Buffer.from(signatureObj.signature),
-    recoveryId: signatureObj.recid
+    signature,
+    recoveryId,
   })
 
   const transactionData = {
